# swagger-auto.json
{
  "swagger": "2.0",
  "info": {
    "title": "MTG Cards Database",
    "description": "<a id='logout' href='http://cse341-mtg-database.herokuapp.com/auth/logout'>Logout</a>",
    "version": "1.0.0"
  },
  "host": "cse341-mtg-database.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "auth-token"
    }
  },
  "paths": {
...
    "/decks/": {
      "get": {
        "tags": [
          "Decks"
        ],
        "summary": "Get all decks",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Decks"
        ],
        "summary": "Add new deck to the database",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deckName": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/decks/{deckName}": {
      "get": {
        "tags": [
          "Decks"
        ],
        "summary": "Get one deck by deck name",
        "description": "",
        "parameters": [
          {
            "name": "deckName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/decks/{deckId}": {
      "put": {
        "tags": [
          "Decks"
        ],
        "summary": "Update existing deck in the database",
        "description": "",
        "parameters": [
          {
            "name": "deckId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "deckName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Decks"
        ],
        "summary": "Delete existing deck from database",
        "description": "",
        "parameters": [
          {
            "name": "deckId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/decks/{deckId}/{cardId}": {
      "post": {
        "tags": [
          "Decks"
        ],
        "summary": "Add card to existing deck",
        "description": "",
        "parameters": [
          {
            "name": "deckId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Decks"
        ],
        "summary": "Remove card from existing deck",
        "description": "",
        "parameters": [
          {
            "name": "deckId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "cardId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/types/": {
      "get": {
        "tags": [
          "Types"
        ],
        "summary": "Get all card types",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Types"
        ],
        "summary": "Add new card type to the database",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "typeId": {
                  "example": "any"
                },
                "typeName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/types/{typeName}": {
      "get": {
        "tags": [
          "Types"
        ],
        "summary": "Get one card type by type name",
        "description": "",
        "parameters": [
          {
            "name": "typeName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Types"
        ],
        "summary": "Update existing card type in databases",
        "description": "",
        "parameters": [
          {
            "name": "typeName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "typeId": {
                  "example": "any"
                },
                "typeName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Types"
        ],
        "summary": "Delete existing card from database",
        "description": "",
        "parameters": [
          {
            "name": "typeName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  }
}